Fixes issue #8 - gcc catch by value and implicit fallthrough warnings
This pr also fixes a number of other minor warning message issues
as documented below.

This pr corrects the majority of both of these warning types.

It also corrects an issue in CIMCLICommand with a type mismatch in
comparison

Update OSINFO.cpp sprintf max size specification

Removed XMLProcess.* throw specification

This is deprecated in C++11 and is to be removed in C++17 and shows up
as warning with gcc. The only goal was to limit the exceptions that
could be thrown by the entity for which it was included, not to throw a
specific exception.

This is no real reason for it in our code.

Modify catch parameter to make it a reference

Add bread to case statements to remove compiler warning

Make catch parameter a reference in TestListener.cpp

Make catch parameter reference in ReliableIndicationDisableEnable test

Increate sprintf buffer size in Process_Linux.cpp to remove warning

Make catch parameter reference in Indication test client

Add break; statemetn to InteropProvider.cpp

Add break; statement to CQLFunctionRep.cpp

Modify catch parameter to define as reference in CQLValue.cpp

Outdent lines in CIMOperationRequestDecoder.cpp that had warning

Add FALLTHROUGH comment to case in PUllErrors.cpp

Separate string and var by space to avoid warning in InternalException.cpp