Fixes issue #7 - deprecated warning message on use of readdir_r
This function is considered obsolete and is deprecated in favor of using
readdir directly.  OpenPegasus uses readdir_r in two ways.

1. Several methods use it to access the disk directories, primarily some
providers.

2. There is a common class in OpenPegasus (src/Pegasus/Common/DirPOSIX
that is available to any Pegasus functionality that uses readdir_r

Since our usage is somewhat complex in that there are threading issues
in moving from readdir_r to readdir, we are simply hiding the warning
for the moment.