Fixes issue # 17 - Modify OpenSSL API
Modifies the OpenSSL api to be compatible with  OpenSSL 1.1.x.

The API changed between Openssl 1.0.2 and 1.1.0 in an incompatible
manner so that we have to update the OpenSSL interface to handle both
the old and new interfaces.

This involves moveing to opqque structures and getter and setter methods
for many of the
interfaces.

NOTE: OpenPegasus must be compiled for one version or the other since
the interfaces are not binary compatible.