Fix several of the prebuilt certificates
Fixes the testchca, testinter certificates so that the include a txt
file to define the subject, a csr and a completed cert.

Adds the bash script create_certs that builds new certs for those certs
that are NOT built as part  of the Makefile.  This allows the certs to
be rebuilt in the future if requirements change.  NOTE: Normally these
certs should not have to be rebuilt unless openssl requirements change
(ex. new minimum key size, etc.)