Fix issue #68 TLSV_1_2_support test fails
Ths failure was an additional incompatibility between OpenPegasus and
OpenSSL.  See the issue for more information.

This pr also modified the name of some internal SSLContext variable s to
the internal name matches the config file and fixes spelling errors
found in the cimconfig help file.

Finally, it updates the changes.rst file to bring it up to date with the
changes to date.