Update Readme file for build and test.
This PR extends the Readme file for the github project and
adds new markdown files in docs file to provide an overview
for build, testing, and running OpenPegasus in
OpenPegasus/docs.BuildOpenPegasusReadme.md and
OpenPegasus/docs/RunningOpenPegasusreadme.md.

This pr also includes a mechanical transformation of the
OpenPegasus pegasus/doc/BuildAndReleaseOptions.html to
markdown format as OpenPegsus/docs/BuildAndReleaseOptions.md
since there appears to be no reasonable way to directly reference
the original document in the github repository.

It also changes the name of the file docs/changes.rst to
changes.md since we will be using Markdown for documentation
in this docs directory.