Add make target for sanity test to OpenPegasus
Adds a new target to the OpenPegasus Makefile, sanitytest which executes
the same target in the TestMakefile.  This target will conduct a very
limited set of tests on OpenPegasus that should take no more that a
minute or two.  This  allows some testing without the extensive unittest
or server tests that can take an hour or more to execute.